chore: :hammer: Rollback webpack for production
