+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "skupper-network-console",
-  "version": "2.0.0-dev",
+  "version": "2.0.0",
   "description": "The Network Console for monitoring Skupper metrics and topology.",
   "repository": "git+https://github.com/skupperproject/skupper-console.git",
   "author": "The Skupper Team",
@@ -17,9 +17,8 @@
     "yarn": ">=1.22 <2.0.0"
   },
   "scripts": {
-    "build": "tsc && vite build",
+    "build": "tsc && webpack --config webpack.config.mjs",
     "build:preview": "vite preview",
-    "build:prod": "tsc && NODE_ENV=production webpack",
     "start": "vite --host",
     "test": "yarn test:watch --run",
     "test:watch": "USE_MOCK_SERVER=true vitest",
@@ -66,10 +65,10 @@
     "@vitejs/plugin-react": "^4.3.4",
     "@vitest/coverage-v8": "^2.1.8",
     "@vitest/eslint-plugin": "^1.1.25",
+    "circular-dependency-plugin": "^5.2.2",
     "commitizen": "^4.3.1",
     "copy-webpack-plugin": "^12.0.2",
     "css-loader": "^7.1.2",
-    "css-minimizer-webpack-plugin": "^7.0.0",
     "cypress": "^13.17.0",
     "cz-git": "^1.11.0",
     "eslint": "^9.18.0",
@@ -89,13 +88,9 @@
     "start-server-and-test": "^2.0.9",
     "terser-webpack-plugin": "^5.3.11",
     "ts-loader": "^9.5.2",
-    "ts-node": "^10.9.2",
     "typescript": "^5.7.3",
     "typescript-eslint": "^8.20.0",
     "vite": "^6.0.7",
-    "vite-plugin-circular-dependency": "^0.5.0",
-    "vite-plugin-favicons-inject": "^2.2.0",
-    "vite-plugin-static-copy": "^2.2.0",
     "vitest": "^2.1.8",
     "webpack": "^5.97.1",
     "webpack-cli": "^6.0.1"
